services/api_service/requirements.txt
# API Service Dependencies
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.3
httpx==0.24.1

# LangChain and LLM
langchain==0.0.267
openai==1.0.0

# Vector DB
pinecone-client==2.2.2
sentence-transformers==2.2.2

# Data processing
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2

# HubSpot Integration
hubspot-api-client==7.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.0
